+++ eslint.config.js
@@ -1,6 +1,4 @@
-// @ts-expect-error
 import eslint from '@eslint/js';
-// @ts-expect-error
 import prettier from 'eslint-plugin-prettier/recommended';
 import ts from 'typescript-eslint';
 
@@ -9,14 +7,31 @@ export default ts.config(
     ignores: ['coverage/', 'dist/', 'docs/', 'tmp/'],
   },
   eslint.configs.recommended,
-  ...ts.configs.strict,
+  ...ts.configs.strictTypeChecked,
+  {
+    languageOptions: {
+      parserOptions: {
+        project: true,
+        tsconfigRootDir: import.meta.dirname,
+      },
+    },
+  },
   prettier,
   {
     rules: {
-      '@typescript-eslint/ban-ts-comment': 'off',
-      '@typescript-eslint/no-explicit-any': 'warn',
+      '@typescript-eslint/no-confusing-void-expression': ['error', { ignoreArrowShorthand: true }],
       '@typescript-eslint/no-invalid-void-type': 'warn',
+      '@typescript-eslint/no-misused-promises': [
+        'error',
+        {
+          checksVoidReturn: {
+            arguments: false,
+          },
+        },
+      ],
       '@typescript-eslint/no-non-null-assertion': 'warn',
+      '@typescript-eslint/related-getter-setter-pairs': 'warn',
+      '@typescript-eslint/restrict-template-expressions': ['error', { allowNumber: true }],
       'no-console': 'error',
       'no-debugger': 'error',
       'prettier/prettier': 'warn',