+++ src/indexeddb/indexeddb.ts
@@ -1,7 +1,7 @@
 /** @module IndexedDB */
 
 import type { RPCClientStrategy } from '../rpc/client/types.js';
-import type { ContentProcedures, PutRequestPayload } from '../rpc/shared/index.js';
+import type { PutRequestPayload } from '../rpc/shared/index.js';
 
 /** Configuration object for the IndexedDB channel driver. */
 export interface IndexedDBConfig {
@@ -25,14 +25,12 @@ export interface IndexedDBConfig {
  * @param config An optional {@linkcode IndexedDBConfig} object to configure aspects of the strategy.
  * @returns An RPC client strategy for IndexedDB.
  */
-export async function indexeddb(
-  config?: IndexedDBConfig,
-): Promise<RPCClientStrategy<ContentProcedures>> {
+export async function indexeddb(config?: IndexedDBConfig): Promise<RPCClientStrategy> {
   const database = config?.database ?? 'astrobase';
   const table = config?.table ?? 'astrobase';
   const db = await new Promise<IDBDatabase>((resolve, reject) => {
     const request = indexedDB.open(database);
-    request.onerror = () => reject(request.error);
+    request.onerror = () => reject(request.error!);
     request.onsuccess = () => resolve(request.result);
     request.onupgradeneeded = () => {
       request.result.createObjectStore(table, { keyPath: 'cid' });
@@ -43,15 +41,16 @@ export async function indexeddb(
       'content:delete': (cid) =>
         new Promise((resolve, reject) => {
           const request = db.transaction(table, 'readwrite').objectStore(table).delete(cid.bytes);
-          request.onerror = () => reject(request.error);
+          request.onerror = () => reject(request.error!);
           request.onsuccess = () => resolve();
         }),
 
       'content:get': (cid) =>
         new Promise((resolve, reject) => {
           const request = db.transaction(table, 'readonly').objectStore(table).get(cid.bytes);
-          request.onerror = () => reject(request.error);
-          request.onsuccess = () => resolve((request.result as { content: ArrayBuffer })?.content);
+          request.onerror = () => reject(request.error!);
+          request.onsuccess = () =>
+            resolve((request.result as { content: ArrayBuffer } | undefined)?.content);
         }),
 
       'content:put': ({ cid, content }: PutRequestPayload) =>
@@ -60,7 +59,7 @@ export async function indexeddb(
             .transaction(table, 'readwrite')
             .objectStore(table)
             .put({ cid: cid.bytes, content });
-          request.onerror = () => reject(request.error);
+          request.onerror = () => reject(request.error!);
           request.onsuccess = () => resolve();
         }),
     },